Refactor Table and Curve to be non-editable
